Atualização...

+++ app/Http/Controllers/ProductController.php
@@ -36,7 +36,7 @@ class ProductController extends Controller
     {
         $dados = $request->all();
 
-        if ($request->hasFile('photo') && $_FILES['photo']['name'] != "padrao.png") {
+        if ($request->hasFile('photo')) {
             $ext = $dados['photo']->extension();
             $imageName = time() . '.' . $ext;
             $dados['photo']->storeAs('produtos', $imageName, 'upload');
@@ -68,17 +68,14 @@ class ProductController extends Controller
 
         if ($product) {
             $dados = $request->all();
-            
-            if ($request->hasFile('photo') && $_FILES['photo']['name'] != "padrao.png") {
+
+            if ($request->hasFile('photo')) {
                 $ext = $dados['photo']->extension();
                 $imageName = time() . '.' . $ext;
                 $dados['photo']->storeAs('produtos', $imageName, 'upload');
                 $dados['photo'] = $imageName;
-            } else {
-                $dados['photo'] = '';
+                Storage::disk('upload')->delete('produtos/' . $product->photo);
             }
-
-            Storage::disk('upload')->delete('produtos/'.$product->photo);
 
             $product->update($dados);
 
@@ -101,15 +98,28 @@ class ProductController extends Controller
                 $movement['total'] -= $movementProduct->pivot->quantity * $movementProduct->pivot->value;
                 $movement->update();
             }
-            Storage::disk('upload')->delete('produtos/'.$product->photo);
+            Storage::disk('upload')->delete('produtos/' . $product->photo);
             $product->delete();
         }
     }
 
-    public function downloadPhoto($id) 
+    public function deletePhoto($id)
     {
         $product = User::find(Auth::id())->products()->find($id);
-        $imageName = $product->name.'_'.$product->photo;
-        return Storage::disk('upload')->download('produtos/'.$product->photo, $imageName);
+        if ($product) {
+            Storage::disk('upload')->delete('produtos/' . $product->photo);
+            $product['photo'] = '';
+            $product->update();
+        }
+    }
+
+    public function downloadPhoto($id)
+    {
+        $product = User::find(Auth::id())->products()->find($id);
+        if (!empty($product->photo)) {
+            $imageName = $product->name . '_' . $product->photo;
+            return Storage::disk('upload')->download('produtos/' . $product->photo, $imageName);
+        } 
+        abort('404');
     }
 }
